requirements2.txt
asgiref                       3.8.1
certifi                       2024.8.30
charset-normalizer            3.4.0
Django                        5.1.1
django-cors-headers           4.4.0
django-filter                 24.3
django-sslserver              0.22
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
idna                          3.10
Markdown                      3.7
marshmallow                   3.17.0
mysqlclient                   2.2.4
packaging                     24.2
pillow                        10.4.0
pip                           24.3.1
Pygments                      2.18.0
PyJWT                         2.9.0
python-decouple               3.8
requests                      2.32.3
setuptools                    63.2.0
sqlparse                      0.5.1
transbank-sdk                 5.0.0
typing_extensions             4.12.2
tzdata                        2024.1
urllib3                       2.2.3